Rename gn3/case_attributes.py -> gn3/api/case_attributes.py.
